+++ CHANGELOG.md
@@ -10,7 +10,9 @@
 
 ### Chores
 
-- Update `cardano-node` to 10.1.2 via `cardano-wallet` v2024-11-18 ([PR 3229](https://github.com/input-output-hk/daedalus/pull/3229))
+- Update `cardano-node` to 10.1.3 ([PR 3249](https://github.com/input-output-hk/daedalus/pull/3249))
+
+- Update `cardano-wallet` to v2024-11-18 ([PR 3229](https://github.com/input-output-hk/daedalus/pull/3229))
 
 ### Fixes